// src/app/models/thread.model.ts
// Debebés: borrar
export class Thread {
    content: string;
    subject: string;
    politicalPartyId: number;
    userId: number
    created?: Date;
    email?: string;
    id?: number;
    password?: string;
    picture?: string;
    postThreadId?: number;
    replies?: number;
    status?: number;
    username?: string;
}